Move the function(SaveUser())
